Модуляция. Исправил ошибки
